Chnages in shader.frag files and values in light, dont work shadows

+++ src/Light.cpp
@@ -1,7 +1,7 @@
 #include "Light.h"
 #include "ShadowMap.h"
 
-Light::Light(GLfloat _width, GLfloat _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity)
+Light::Light(GLuint _width, GLuint _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity)
         : colour(_colour), ambientIntensity(_aIntensity), diffuseIntensity(_dIntensity) {
     shadowMap = new ShadowMap();
     shadowMap->Init(_width, _height);
